exportStories だけにして
