+++ .gitignore
@@ -1,5 +1,3 @@
 cdk.out/
 dist/
 node_modules/
-s3/hacker-mainichi/*
-!s3/hacker-mainichi/.gitkeep
