+++ .prettierignore
@@ -1,3 +1,2 @@
 cdk.out/
 dist/
-s3/
