+++ cdk.json
@@ -1,3 +1,3 @@
 {
-  "app": "npx ts-node ./src"
+  "app": "npx ts-node ./src/api.ts"
 }
