+++ src/lib/index.ts
@@ -1,3 +1,2 @@
 export * from './api';
 export * from './db';
-export * from './storage';
