+++ tsconfig.json
@@ -3,10 +3,11 @@
     "esModuleInterop": true,
     "incremental": true,
     "isolatedModules": true,
+    "lib": ["ES2019"],
     "module": "CommonJS",
     "sourceMap": true,
     "strict": true,
-    "target": "ESNext",
+    "target": "ES2019",
     "outDir": "./dist"
   },
   "include": ["./src"]
